[PTQ] MinMax algo improvements
